Move keytar-strategy to lib/shared/
