+++ lib/controllers/issueish-pane-item-controller.js
@@ -6,7 +6,7 @@ import {QueryRenderer, graphql} from 'react-relay';
 
 import RelayNetworkLayerManager from '../relay-network-layer-manager';
 import GithubLoginModel from '../models/github-login-model';
-import {UNAUTHENTICATED} from '../models/keytar-strategy';
+import {UNAUTHENTICATED} from '../shared/keytar-strategy';
 import GithubLoginView from '../views/github-login-view';
 import ObserveModel from '../views/observe-model';
 import IssueishLookupByNumberContainer from '../containers/issueish-lookup-by-number-container';
