+++ lib/controllers/pr-info-controller.js
@@ -8,7 +8,7 @@ import PrSelectionByUrlContainer from '../containers/pr-selection-by-url-contain
 import PrSelectionByBranchContainer from '../containers/pr-selection-by-branch-container';
 import GithubLoginView from '../views/github-login-view';
 import RelayNetworkLayerManager from '../relay-network-layer-manager';
-import {UNAUTHENTICATED} from '../models/keytar-strategy';
+import {UNAUTHENTICATED} from '../shared/keytar-strategy';
 
 export default class PrInfoController extends React.Component {
   static propTypes = {
