+++ lib/controllers/remote-pr-controller.js
@@ -6,7 +6,7 @@ import yubikiri from 'yubikiri';
 import {RemotePropType} from '../prop-types';
 import ObserveModelDecorator from '../decorators/observe-model';
 import GithubLoginView from '../views/github-login-view';
-import {UNAUTHENTICATED} from '../models/keytar-strategy';
+import {UNAUTHENTICATED} from '../shared/keytar-strategy';
 import {nullRemote} from '../models/remote';
 import PrInfoController from './pr-info-controller';
