+++ lib/models/github-login-model.js
@@ -1,6 +1,6 @@
 import {Emitter} from 'event-kit';
 
-import {UNAUTHENTICATED, createStrategy} from './keytar-strategy';
+import {UNAUTHENTICATED, createStrategy} from '../shared/keytar-strategy';
 
 let instance = null;
