+++ lib/shared/keytar-strategy.js
@@ -1,6 +1,5 @@
-import {execFile} from 'child_process';
-import path from 'path';
-import fs from 'fs';
+const {execFile} = require('child_process');
+const fs = require('fs');
 
 if (typeof atom === 'undefined') {
   global.atom = {
@@ -9,9 +8,9 @@ if (typeof atom === 'undefined') {
   };
 }
 
-export const UNAUTHENTICATED = Symbol('UNAUTHENTICATED');
+const UNAUTHENTICATED = Symbol('UNAUTHENTICATED');
 
-export class KeytarStrategy {
+class KeytarStrategy {
   static get keytar() {
     return require('keytar');
   }
@@ -49,7 +48,7 @@ export class KeytarStrategy {
   }
 }
 
-export class SecurityBinaryStrategy {
+class SecurityBinaryStrategy {
   static isValid() {
     return process.platform === 'darwin';
   }
@@ -81,7 +80,7 @@ export class SecurityBinaryStrategy {
   }
 }
 
-export class InMemoryStrategy {
+class InMemoryStrategy {
   static isValid() {
     return true;
   }
@@ -117,7 +116,7 @@ export class InMemoryStrategy {
   }
 }
 
-export class FileStrategy {
+class FileStrategy {
   static isValid() {
     if (!atom.inSpecMode() && !atom.inDevMode()) {
       return false;
@@ -216,7 +215,7 @@ export class FileStrategy {
 const strategies = [FileStrategy, KeytarStrategy, SecurityBinaryStrategy, InMemoryStrategy];
 let ValidStrategy = null;
 
-export async function createStrategy() {
+async function createStrategy() {
   if (ValidStrategy) {
     return new ValidStrategy();
   }
@@ -234,3 +233,12 @@ export async function createStrategy() {
   }
   return new ValidStrategy();
 }
+
+module.exports = {
+  UNAUTHENTICATED,
+  KeytarStrategy,
+  SecurityBinaryStrategy,
+  InMemoryStrategy,
+  FileStrategy,
+  createStrategy,
+};
